fix(listsection): consolidating ListSectionGroup to ListSection
