--- packages/patterns-react/src/patterns/ListSection/ListSectionGroup.js
@@ -1,67 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { ContentGroup } from '@carbon/ibmdotcom-react';
-import ListSectonItem from './ListSectionItem';
-import PropTypes from 'prop-types';
-import React from 'react';
-import { settings as ddsSettings } from '@carbon/ibmdotcom-utilities';
-import { settings } from 'carbon-components';
-
-const { stablePrefix } = ddsSettings;
-const { prefix } = settings;
-
-/**
- * List Section Group Component
- *
- * @param {object} props props object
- * @param {string} props.title List Group title
- * @param {Array} props.lists link Group lists array
- * @returns {object} JSX Object
- */
-const ListSectionGroup = ({ listGroup: { title, lists } }) => {
-  return (
-    <>
-      {/* <div
-      data-autoid={`${stablePrefix}--listsection-group`}
-      className={`${prefix}--listsection-group`}>
-      <h2 className={`${prefix}--listsection-group__title`}>{title}</h2>
-      <div className={`${prefix}--listsection-group__list`}>
-        {_renderList(lists)}
-      </div>
-    </div> */}
-      <ContentGroup
-        className={`${prefix}--listsection-group`}
-        data-autoid={`${stablePrefix}--listsection-group`}
-        heading={title}>
-        {_renderList(lists)}
-      </ContentGroup>
-    </>
-  );
-};
-
-/**
- * Render List Component
- *
- * @private
- * @param {object} listItems listItems Object
- * @returns {object} JSX Object
- */
-const _renderList = listItems => {
-  return listItems.map(listItem => {
-    return <ListSectonItem key={listItem.title} lists={listItem} />;
-  });
-};
-
-ListSectionGroup.propTypes = {
-  listGroup: PropTypes.shape({
-    title: PropTypes.string,
-    lists: PropTypes.array,
-  }),
-};
-
-export default ListSectionGroup;
